requirements.txt
amqp==2.5.2
anyjson==0.3.3
asn1crypto==0.24.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
Django==2.2.7
django-mailgun==0.9.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
freezegun==0.3.12
httplib2==0.9.2
idna==2.8
importlib-metadata==0.23
itypes==1.1.0
Jinja2==2.10.3
keyring==10.6.0
keyrings.alt==3.0
kombu==4.6.6
launchpadlib==1.10.7
lazr.restfulclient==0.14.2
macaroonbakery==1.2.3
Mako==1.0.7
MarkupSafe==1.1.1
more-itertools==7.2.0
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
openapi-codec==1.3.2
pexpect==4.2.1
Pillow==6.2.1
protobuf==3.0.0
psycopg2==2.8.4
pycairo==1.16.2
pycrypto==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
pyxdg==0.25
PyYAML==5.1.2
redis==3.3.11
reportlab==3.4.0
requests==2.22.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.16.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
Werkzeug==0.16.0
zipp==0.6.0
